# rafs: prefetch based on blob chunks rather than files