@changweige 2022-11-29
> Please also open a PR for master? It should also improve master branch:)

Sure. :-) Will raise another PR to master branch